requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-pptx==1.0.2
openai
python-dotenv